gemini-pdf-summarizer/requirements.txt
flask>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
uuid>=1.30